Rename manifest-compare files for naming consistency
Mirror the manifest-generate convention where the file/function named
after the mode is the top-level mode entry point.

- manifest-compare.ts (classifier) -> manifest-compare-classify.ts
  (manifestCompare -> classifyManifests; ManifestCompareDeps ->
  ClassifyDeps; CompareParams -> ClassifyParams)
- run-manifest-compare.ts (orchestrator) -> manifest-compare.ts
  (runManifestCompare -> manifestCompare; RunManifestCompareDeps ->
  ManifestCompareDeps; RunManifestCompareParams -> ManifestCompareParams)
